requirements.txt
llama-index
python-dotenv
openai
pypdf